msm8998: Stage 1 change to move the dsp partition into A/B mode
This change will try and mount both dsp and dsp_a partitions. This
is the first step in order to move the dsp partition to A/B mode, and
is required so that the build does not break when the partition table
is updated.

Once the updated partition table with dsp_a/b merges, we will move
the mount step for dsp into the fstab.

Change-Id: Icc020d5238beea926ede8944d8a147632333e288